% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalnexus.R
\name{animalnexus}
\alias{animalnexus}
\alias{animalnexus-site}
\title{Launch local animalnexus}
\usage{
animalnexus()
}
\description{
A local shiny app of the animalnexus site allowing users to import, transform and visualize data locally. Note that visualizations require an internet connection to download map tiles.
}
\examples{

\dontrun{
animalnexus()
}

}
