% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_data.R
\name{dl_data}
\alias{dl_data}
\title{Download data from original animalnexus database}
\usage{
dl_data(
  start = NULL,
  end = NULL,
  url = "http://gaia.tru.ca/birdMOVES/rscripts/anquery.csv",
  tz_disp = "Etc/GMT+8",
  dst = FALSE,
  species = NULL,
  site_id = NULL,
  feeder_details,
  bird_details,
  logger_details,
  animal_details
)
}
\arguments{
\item{start}{Character. This is the start date (inclusive, with or without
time) for the data to download. There is some flexibility in the format
(see details). If NULL, get records from start.}

\item{end}{Character. This is the end date (inclusive, with or without time)
for the data to download. There is some flexibility in the format (see
details). If NULL, get records to end.}

\item{url}{Character. This is the url for the database service. The default
should not need to be changed.}

\item{tz_disp}{Character vector. Timezone data should be displayed in (should match one of
the zones produced by `OlsonNames()`)}

\item{dst}{Logical. Whether or not to use Daylight Savings. When set to FALSE
timezones are converted to the Etc/GMT+X timezones which do not include
DST. (Note this overrides the timezone specification such that a timezone
of America/Vancouver, which would normally include DST in the summer, will
be transformed to a timezone with the same GMT offset, but not including
DST).}

\item{species}{Character. Vector of species to include (defaults to all). See
details for valid entries.}

\item{site_id}{Character. Vector of sites to include (defaults to all
permissible). See details for valid entries.}

\item{feeder_details}{Deprecated.}

\item{bird_details}{Deprecated.}

\item{logger_details}{Deprecated.}

\item{animal_details}{Deprecated.}
}
\description{
This function uses RCurl to submit an HTML form and retrieve the csv file
from the animalnexus database.
}
\details{
This function is flexible with respect to date formats. Using
[lubridate::parse_date_time()] from the lubridate package, the
format of the date/time supplied will estimated. This allows for partial
date/times to be supplied (e.g., "2015-01-01 09" or "2015-09" or
"2015-09-01"). For best results, specify the date/time format as YYYY-MM-DD
HH:MM:SS and truncate as needed. Note that truncated times are interpreted as
00 and trucated dates as the first of the month and the first month of the
year. Thus "2015" will be sumbitted as "2015-01-01 00:00:00".

`species` options are:
\itemize{
  \item House Finch
  \item Mountain Chickadee
  \item Dark-eyed Junco
  \item Green Hermit
  \item Violet Sabrewing
  \item Rufous-tailed Hummingbird
  \item Stripe-throated Hermit
  }

`site_id` options are either 'kl' for Kamloops, BC, or 'cr' for Costa
Rica. Note that Costa Rican data is protected while the scientist work on
publishing. Only users with valid credentials will be able to download this
data.
}
\examples{
\dontrun{

# Get all data (may take a couple minutes)
r <- dl_data()

# Get all 2016 data
r <- dl_date(start = "2016")

# Get specific data
r <- dl_data(start = "2016-01-01 09:34:12",
              end = "2016-02-01")
}

}
