% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedrUI-pkg.R
\docType{package}
\name{feedr-package}
\alias{feedr-package}
\alias{feedrUI}
\alias{feedrUI-package}
\title{Shiny UI for feedr package}
\description{
User-friendly Shiny user interfaces for working with the feedr package.
}
\references{
animalnexus <http://animalnexus.ca>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/animalnexus/feedrUI}
}

}
\author{
\strong{Maintainer}: LaZerte Steffi \email{sel@steffilazerte.ca}

}
