% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_animate.R
\name{ui_animate}
\alias{ui_animate}
\title{Animated map for individuals with leaflet}
\usage{
ui_animate(v, verbose = FALSE)
}
\arguments{
\item{v}{Data frame. Visits data frame created with the \code{\link{visits}}
function.}

\item{verbose}{Logical. Print log events to console.}
}
\description{
Interactive shiny app to select and animate presence at and movements between
RFID loggers over time. Also available online at \url{http://animalnexus.ca} or
by launching the local animalnexus app through \code{\link{animalnexus}}.
}
\examples{

\dontrun{
map_animate_indiv(visits(finches))
}

}
