% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_db.R
\name{ui_db}
\alias{ui_db}
\title{User-interface for downloading from the animalnexus database}
\usage{
ui_db(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Print log events to console.}
}
\value{
Downloaded data frame formatted and ready to be transformed
}
\description{
Launches an interactive shiny app for downloading data interactively. Also
available online at \url{http://animalnexus.ca} or by launching the local
animalnexus app through \code{\link{animalnexus}}. See the \code{\link{dl_data}}
function for a non-interactive method.
}
\examples{

\dontrun{
  my_data <- ui_db()
}

}
\seealso{
\code{\link{dl_data}}
}
