% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trans.R
\name{ui_trans}
\alias{ui_trans}
\title{User-interface for transforming data}
\usage{
ui_trans(r, verbose = FALSE)
}
\arguments{
\item{r}{Data frame. Raw RFID data to transform}

\item{verbose}{Logical. Print log events to console.}
}
\value{
Returns nothing. Data can be saved to disk from the user-interface,
  but transformation are not returned to the current R session.
}
\description{
An interactive shiny app for transforming data. Also available online at
<http://animalnexus.ca> or by launching the local animalnexus app through
\code{animalnexus()}.
}
\details{
See individual data transformations for more details: \code{visits()},
\code{presence()}, \code{move()}, \code{disp()}, \code{activity()}, and
\code{daily()}.
}
\examples{
\dontrun{
  ui_trans(r = my_data)
}

}
